Working example

+++ Sokoban/Utilities/Extensions/TypeConversions.cs
@@ -6,9 +6,9 @@ namespace Sokoban.Utilities.Extensions
 {
 public static partial class Extension
 {
-  public static Vector2D<double> ToVector2D(this ScrollWheel wheel) => new(wheel.X, wheel.Y);
-  public static Vector2D<double> ToVector2D(this Vector2 vector) => new(vector.X, vector.Y);
-  public static Vector3D<double> ToVector3D(this Vector3 vector) => new(vector.X, vector.Y, vector.Z);
-  public static Vector4D<double> ToVector4D(this Vector4 vector) => new(vector.X, vector.Y, vector.Z, vector.W);
+  public static Vector2D<float> ToVector2D(this ScrollWheel wheel) => new(wheel.X, wheel.Y);
+  public static Vector2D<float> ToVector2D(this Vector2 vector) => new(vector.X, vector.Y);
+  public static Vector3D<float> ToVector3D(this Vector3 vector) => new(vector.X, vector.Y, vector.Z);
+  public static Vector4D<float> ToVector4D(this Vector4 vector) => new(vector.X, vector.Y, vector.Z, vector.W);
 }
 }
